[DR-3168] BucketResourceTest should use globally-unique bucket names (#1493)
* [DR-3168] BucketResourceTest should use globally-unique bucket names

GCS requires that bucket names be globally unique.  Previously, this test suite reused fixed bucket names.
When buckets were properly removed on test tear-down and the same test was not run concurrently, this didn't yield issues.
But recently we think that a bucket was not properly torn down, and its parent project was marked for deletion, but per GCS documentation this means that the bucket may stick around for weeks before being fully removed.
To guard against this issue, any bucket names created in this test are now guaranteed to be globally unique.  We also try to ensure their adherence to other GCS bucket naming requirements (length, allowed characters, etc.) via an assertion, so that future debugging is hopefully made easier.

Future work: pursue opportunities to refactor some of these tests into unit tests.

* @pshapiro4broad PR feedback

+++ src/test/java/bio/terra/service/resourcemanagement/google/BucketResourceTest.java
@@ -4,6 +4,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.equalToIgnoringCase;
 import static org.hamcrest.Matchers.hasItem;
+import static org.hamcrest.Matchers.matchesPattern;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
@@ -88,7 +89,6 @@ public class BucketResourceTest {
   private BillingProfileModel profile;
   private Storage storage;
   private List<GoogleBucketResource> bucketResources;
-  private boolean allowReuseExistingBuckets;
   private GoogleProjectResource projectResource;
   private UUID datasetId;
 
@@ -121,9 +121,7 @@ public class BucketResourceTest {
 
   @After
   public void teardown() throws Exception {
-    for (GoogleBucketResource bucketResource : bucketResources) {
-      deleteBucket(bucketResource);
-    }
+    bucketResources.forEach(this::deleteBucket);
     // Connected operations resets the configuration
     connectedOperations.teardown();
   }
@@ -131,7 +129,7 @@ public class BucketResourceTest {
   @Test
   // create and delete the bucket, checking that the metadata and cloud state match what is expected
   public void createAndDeleteBucketTest() throws Exception {
-    String bucketName = "testbucket_createanddeletebuckettest";
+    String bucketName = createBucketName("createanddeletebuckettest");
     String flightId = "createAndDeleteBucketTest";
 
     // create the bucket and metadata
@@ -158,12 +156,11 @@ public class BucketResourceTest {
   // create buckets in different regions and see if they're actually created there.
   public void createBucketsInDifferentRegionsTest() throws Exception {
     for (GoogleRegion region : List.of(GoogleRegion.US_CENTRAL1, GoogleRegion.US_EAST1)) {
-      String bucketName = "testbucket_bucketregionstest_" + region;
+      String bucketName = createBucketName(region.toString());
       String flightId = "bucketRegionsTest_" + region;
 
       // create the bucket and metadata
-      GoogleBucketResource bucketResource =
-          createBucket(bucketName, projectResource, region, flightId, null, null, null, true);
+      createBucket(bucketName, projectResource, region, flightId, null, null, null, true);
 
       // Get the Bucket
       Bucket cloudBucket = bucketService.getCloudBucket(bucketName);
@@ -177,7 +174,7 @@ public class BucketResourceTest {
 
   @Test
   public void testMultiRegionalBucket() throws InterruptedException {
-    String bucketName = "testbucket_multiregiontest";
+    String bucketName = createBucketName("multiregiontest");
     String flightId = "testMultiRegionalBucket";
 
     // create the bucket and metadata
@@ -204,7 +201,7 @@ public class BucketResourceTest {
   // after it's been created, confirm it succeeds.
   public void twoThreadsCompeteForLockTest() throws Exception {
     String flightIdBase = "twoThreadsCompeteForLockTest";
-    String bucketName = "twothreadscompeteforlocktest";
+    String bucketName = createBucketName("twothreadscompetetest");
     GoogleRegion bucketRegion = GoogleRegion.DEFAULT_GOOGLE_REGION;
 
     BucketResourceLockTester resourceLockA =
@@ -278,7 +275,7 @@ public class BucketResourceTest {
         "property allowReuseExistingBuckets = {}",
         bucketResourceUtils.getAllowReuseExistingBuckets(configService));
 
-    String bucketName = "testbucket_bucketexistsbeforemetadatatest";
+    String bucketName = createBucketName("bucketbeforemetadatatest");
     String flightIdA = "bucketExistsBeforeMetadataTestA";
 
     // create the bucket and metadata
@@ -359,7 +356,7 @@ public class BucketResourceTest {
         "property allowReuseExistingBuckets = {}",
         bucketResourceUtils.getAllowReuseExistingBuckets(configService));
 
-    String bucketName = "testbucket_nobucketbutmetadataexiststest";
+    String bucketName = createBucketName("nobucketbutmetadatatest");
     String flightIdA = "noBucketButMetadataExistsTestA";
 
     // create the bucket and metadata
@@ -420,10 +417,10 @@ public class BucketResourceTest {
 
   @Test
   public void createBucketWithTtlTest() throws Exception {
-    String bucketWithTtlName = "testbucket_bucketwithttl";
+    String bucketWithTtlName = createBucketName("bucketwithttltest");
     String flightWithTtlId = "bucketwithttltest";
 
-    String bucketWithoutTtlName = "testbucket_bucketwithoutttl";
+    String bucketWithoutTtlName = createBucketName("bucketwithoutttltest");
     String flightWithoutTtlId = "bucketwithoutttltest";
 
     Integer deleteAge = 1;
@@ -479,7 +476,7 @@ public class BucketResourceTest {
 
   @Test
   public void testCreateBucketWithReaders() throws Exception {
-    String bucketName = "bucket_with_reader_groups";
+    String bucketName = createBucketName("bucketwithreadergroupstest");
     String flightId = "bucketWithReaderGroupsTest";
     List<String> readerGroups =
         new ArrayList<>(
@@ -520,7 +517,6 @@ public class BucketResourceTest {
       String dedicatedServiceAccount,
       boolean autoclassEnabled)
       throws InterruptedException {
-
     GoogleBucketResource bucketResource =
         bucketService.getOrCreateBucket(
             bucketName,
@@ -550,10 +546,12 @@ public class BucketResourceTest {
   }
 
   private void deleteBucket(GoogleBucketResource bucketResource) {
-    // delete the bucket and update the metadata
-    storage.delete(bucketResource.getName());
-    bucketService.updateBucketMetadata(bucketResource.getName(), null);
-    datasetBucketDao.deleteDatasetBucketLink(datasetId, bucketResource.getResourceId());
+    if (bucketResource != null) {
+      // delete the bucket and update the metadata
+      storage.delete(bucketResource.getName());
+      bucketService.updateBucketMetadata(bucketResource.getName(), null);
+      datasetBucketDao.deleteDatasetBucketLink(datasetId, bucketResource.getResourceId());
+    }
   }
 
   private void checkBucketDeleted(String bucketName, UUID bucketResourceId) {
@@ -590,5 +588,20 @@ public class BucketResourceTest {
         GoogleRegion.DEFAULT_GOOGLE_REGION,
         Map.of("test-name", "bucket-resource-test"),
         CollectionType.DATASET);
+  }
+
+  /**
+   * @param baseName the desired prefix for a GCS bucket name
+   * @return a globally-unique GCS bucket name based off of the provided baseName
+   */
+  private String createBucketName(String baseName) {
+    String globallyUniqueName = Names.randomizeName(baseName.toLowerCase());
+    // For more information on GCS naming requirements:
+    // https://cloud.google.com/storage/docs/buckets#naming
+    assertThat(
+        "Bucket name follows GCS naming requirements",
+        globallyUniqueName,
+        matchesPattern("^[a-z0-9][a-z0-9\\-_]{1,61}[a-z0-9]$"));
+    return globallyUniqueName;
   }
 }
